+++ src/serializer/SerializerRenderer.ts
@@ -4,7 +4,6 @@ import CalendarDate from '../helpers/CalendarDate.js'
 import compact from '../helpers/compact.js'
 import { DateTime } from '../helpers/DateTime.js'
 import round from '../helpers/round.js'
-import sanitizeString from '../helpers/sanitizeString.js'
 import snakeify from '../helpers/snakeify.js'
 import { ViewModel } from '../types/dream.js'
 import {
@@ -71,9 +70,7 @@ export default class SerializerRenderer {
           const outputAttributeName = this.setCase(attribute.options?.as ?? attribute.name)
           const value = data[attribute.name] ?? attribute.options?.default
           accumulator[outputAttributeName] = applyRenderingOptionsToAttribute(
-            attribute.options?.sanitize && typeof value === 'string'
-              ? sanitizeString(value, attribute.options.sanitize)
-              : value,
+            value,
             attribute.options,
             this.passthroughData,
             this.renderOpts
@@ -93,9 +90,7 @@ export default class SerializerRenderer {
           const target = data[attribute.targetName]
           const value = target?.[attribute.name] ?? attribute.options?.default
           accumulator[outputAttributeName] = applyRenderingOptionsToAttribute(
-            attribute.options?.sanitize && typeof value === 'string'
-              ? sanitizeString(value, attribute.options.sanitize)
-              : value,
+            value,
             attribute.options,
             this.passthroughData,
             this.renderOpts